[#23] fail properly for undefined values during generation
