+++ k8t/util.py
@@ -7,32 +7,16 @@
 #
 # THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
-import base64
 import copy
-import hashlib
 import logging
 import os
-import string
 from functools import reduce
-from typing import Any, Dict, List
+from typing import Dict, List
 
 import yaml
 from simple_tools.interaction import confirm
 
 LOGGER = logging.getLogger(__name__)
-
-try:
-    from secrets import choice
-except ImportError:
-    from random import SystemRandom
-
-    choice = SystemRandom().choice
-
-
-def random_password(length: int) -> str:
-    return "".join(
-        choice(string.ascii_lowercase + string.digits) for _ in range(length)
-    )
 
 
 # def include_file(name: str):
@@ -42,48 +26,6 @@ def random_password(length: int) -> str:
 #         return s.read()
 
 
-def b64encode(value: Any) -> str:
-    result = None
-
-    if isinstance(value, str):
-        result = base64.b64encode(value.encode()).decode()
-    elif isinstance(value, int):
-        result = base64.b64encode(str(value).encode()).decode()
-    elif isinstance(value, bytes):
-        result = base64.b64encode(value).decode()
-    else:
-        raise TypeError("invalid input: {}".format(value))
-
-    return result
-
-
-def b64decode(value: Any) -> str:
-    result = None
-
-    if isinstance(value, str):
-        result = base64.b64decode(value.encode()).decode()
-    elif isinstance(value, bytes):
-        result = base64.b64decode(value).decode()
-    else:
-        raise TypeError("invalid input: {}".format(value))
-
-    return result
-
-
-def hashf(value, method="sha256"):
-    try:
-        hash_method = getattr(hashlib, method)()
-    except AttributeError:
-        raise RuntimeError("No such hash method: {}".format(method))
-
-    if isinstance(value, str):
-        hash_method.update(value.encode())
-    elif isinstance(value, bytes):
-        hash_method.update(value)
-    else:
-        raise TypeError("invalid input: {}".format(value))
-
-    return hash_method.hexdigest()
 
 
 def touch(fname: str, mode=0o666, dir_fd=None, **kwargs) -> None:
@@ -159,9 +101,6 @@ def load_yaml(path: str) -> dict:
         return yaml.safe_load(stream) or dict()
 
 
-def envvar(key: str, default=None) -> str:
-    return os.environ.get(key, default)
-
 
 def envvalues() -> Dict:
     prefix: str = 'K8T_VALUE_'
@@ -174,14 +113,14 @@ def envvalues() -> Dict:
     return values
 
 
-def list_files(directory: str, files=False, directories=False) -> List[str]:
+def list_files(directory: str, include_files=False, include_directories=False) -> List[str]:
     result = []
 
     for _, dirs, files in os.walk(directory):
-        if files:
+        if include_files:
             result.extend(files)
 
-        if directories:
+        if include_directories:
             result.extend(dirs)
 
         break
