+++ tests/util.py
@@ -9,8 +9,7 @@
 
 import random
 
-from k8t.util import (b64decode, b64encode, deep_merge, hashf, merge,
-                      random_password)
+from k8t.util import deep_merge, merge
 
 
 def test_merge_memory_safety():
@@ -48,27 +47,3 @@ def test_deep_merge():
 
     assert (
         deep_merge(dict_c, dict_b, dict_a) == dict(foo=dict(a=1, b=2), baz=4, bar=dict(a=3, c=9)))
-
-
-def test_b64encode():
-    string = "foobar"
-
-    encoded = b64encode(string)
-
-    assert encoded != string
-    assert b64decode(encoded) == string
-
-
-def test_random_password():
-    length = int(random.uniform(1, 200))
-
-    assert len(random_password(length)) == length
-    assert random_password(length) != random_password(length)
-
-
-def test_hashf():
-    string = "foobar"
-
-    assert hashf(string) != string
-    assert hashf(string) == hashf(string)
-    assert hashf(string) != hashf("foobaz")
